Add more supported zones and move to test utils
