+++ pubsublite/ps/integration_test.go
@@ -17,7 +17,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math/rand"
 	"strings"
 	"sync/atomic"
 	"testing"
@@ -40,23 +39,7 @@ const (
 	defaultTestTimeout = 5 * time.Minute
 )
 
-var (
-	resourceIDs = uid.NewSpace("go-ps-test", nil)
-	rng         *rand.Rand
-
-	// A random zone is selected for each integration test run.
-	supportedZones = []string{
-		"us-central1-a",
-		"us-central1-b",
-		"us-central1-c",
-		"europe-west1-b",
-		"europe-west1-d",
-	}
-)
-
-func init() {
-	rng = testutil.NewRand(time.Now())
-}
+var resourceIDs = uid.NewSpace("go-ps-test", nil)
 
 func initIntegrationTest(t *testing.T) {
 	if testing.Short() {
@@ -112,15 +95,11 @@ func subscriberClient(ctx context.Context, t *testing.T, settings ReceiveSetting
 	return sub
 }
 
-func randomLiteZone() string {
-	return supportedZones[rng.Intn(len(supportedZones))]
-}
-
 func initResourcePaths(t *testing.T) (string, pubsublite.TopicPath, pubsublite.SubscriptionPath) {
 	initIntegrationTest(t)
 
 	proj := testutil.ProjID()
-	zone := randomLiteZone()
+	zone := test.RandomLiteZone()
 	region, _ := pubsublite.ZoneToRegion(zone)
 	resourceID := resourceIDs.New()
 
@@ -174,6 +153,14 @@ func cleanUpSubscription(ctx context.Context, t *testing.T, admin *pubsublite.Ad
 	} else {
 		t.Logf("Deleted subscription %s", subscription)
 	}
+}
+
+func partitionNumbers(partitionCount int) []int {
+	var partitions []int
+	for i := 0; i < partitionCount; i++ {
+		partitions = append(partitions, i)
+	}
+	return partitions
 }
 
 func publishMessages(t *testing.T, settings PublishSettings, topic pubsublite.TopicPath, msgs ...*pubsub.Message) {
@@ -292,7 +279,7 @@ func TestIntegration_PublishSubscribeSinglePartition(t *testing.T) {
 	ctx := context.Background()
 	const partitionCount = 1
 	recvSettings := DefaultReceiveSettings
-	recvSettings.Partitions = []int{0}
+	recvSettings.Partitions = partitionNumbers(partitionCount)
 
 	admin := adminClient(ctx, t, region)
 	defer admin.Close()
@@ -398,7 +385,7 @@ func TestIntegration_PublishSubscribeSinglePartition(t *testing.T) {
 	})
 
 	// Verifies that SubscriberClient.Receive() can be invoked multiple times
-	// serially (not in parallel).
+	// serially (note: parallel would error).
 	t.Run("SubscriberMultipleReceive", func(t *testing.T) {
 		msgs := []*pubsub.Message{
 			{Data: []byte("multiple_receive1")},
@@ -465,8 +452,8 @@ func TestIntegration_PublishSubscribeSinglePartition(t *testing.T) {
 			}
 
 			// The commit offset for this ack should be processed since the subscriber
-			// not shut down due to fatal error. Not actually detected until the next
-			// test, which would receive an incorrect message.
+			// is not shut down due to fatal error. Not actually detected until the
+			// next test, which would receive an incorrect message.
 			got.Ack()
 		}
 		subscriber := subscriberClient(cctx, t, recvSettings, subscriptionPath)
@@ -534,9 +521,7 @@ func TestIntegration_PublishSubscribeMultiPartition(t *testing.T) {
 	region, topicPath, subscriptionPath := initResourcePaths(t)
 	ctx := context.Background()
 	recvSettings := DefaultReceiveSettings
-	for i := 0; i < partitionCount; i++ {
-		recvSettings.Partitions = append(recvSettings.Partitions, i)
-	}
+	recvSettings.Partitions = partitionNumbers(partitionCount)
 
 	admin := adminClient(ctx, t, region)
 	defer admin.Close()
@@ -552,7 +537,7 @@ func TestIntegration_PublishSubscribeMultiPartition(t *testing.T) {
 
 	// Tests messages published without ordering key.
 	t.Run("PublishRoutingNoKey", func(t *testing.T) {
-		const messageCount = 10 * partitionCount
+		const messageCount = 50 * partitionCount
 
 		// Publish messages.
 		msgs := publishPrefixedMessages(t, DefaultPublishSettings, topicPath, "routing_no_key", messageCount)
@@ -567,7 +552,7 @@ func TestIntegration_PublishSubscribeMultiPartition(t *testing.T) {
 	// Tests messages published with ordering key.
 	t.Run("PublishRoutingWithKey", func(t *testing.T) {
 		const messageCountPerPartition = 100
-		const publishBatchSize = 10 // Verifies ordering of batches
+		const publishBatchSize = 5 // Verifies ordering of batches
 
 		// Publish messages.
 		orderingSender := test.NewOrderingSender()
@@ -596,7 +581,7 @@ func TestIntegration_PublishSubscribeMultiPartition(t *testing.T) {
 	// Verifies usage of the partition assignment service.
 	t.Run("PartitionAssignment", func(t *testing.T) {
 		const messageCount = 100
-		const subscriberCount = 2 // Should be between (1, partitionCount]
+		const subscriberCount = 2 // Should be between [2, partitionCount]
 
 		// Publish messages.
 		msgs := publishPrefixedMessages(t, DefaultPublishSettings, topicPath, "partition_assignment", messageCount)
@@ -661,13 +646,11 @@ func TestIntegration_SubscribeFanOut(t *testing.T) {
 
 	// Receive messages from multiple subscriptions.
 	recvSettings := DefaultReceiveSettings
-	for i := 0; i < partitionCount; i++ {
-		recvSettings.Partitions = append(recvSettings.Partitions, i)
-	}
+	recvSettings.Partitions = partitionNumbers(partitionCount)
 
 	for _, subscription := range subscriptionPaths {
 		msgTracker := test.NewMsgTracker()
 		msgTracker.Add(msgs...)
-		receiveAllMessages(t, msgTracker, recvSettings, subscription, checkOrdering(false))
+		receiveAllMessages(t, msgTracker, recvSettings, subscription, checkOrdering(partitionCount == 1))
 	}
 }
